% Generated by roxygen2 (4.0.2): do not edit by hand
\name{api_dict}
\alias{api.dict}
\alias{api_dict}
\title{API dictionary}
\usage{
api_dict()
}
\description{
By default provided to \code{getOption("Rbitcoin.api.dict")}. This function returns built-in Rbitcoin data set contains API dictionary for \code{\link{market.api.process}} function to perform pre-process API call arguments, post-process API call results and catch market level errors. Still there is function \code{\link{market.api.query}} that do not require any dictionary and can operate on any currency pair. Granularity of api dictionary data is \code{c("market", "base", "quote", "action")}. This dictionary can be edited/extended by user for new markets, currency pairs and actions.\cr Currently supported currency pairs:
\itemize{
\item \code{bitstamp: BTCUSD}
\item \code{btce: BTCUSD, LTCUSD, LTCBTC, NMCBTC}
\item \code{kraken: BTCEUR, LTCEUR, BTCLTC}
\item \code{bitmarket: BTCPLN, LTCPLN}
}
}
\note{
Do not use \code{api.dict} from untrusted source or read whole it's code to ensure it is safe! The api dictionary was not fully tested, please follow the examples, if you find any bugs please report.
}

