% Generated by roxygen2: do not edit by hand
\name{blockchain.api.process}
\alias{blockchain.api.process}
\title{Process blockchain.info API}
\usage{
blockchain.api.process(..., method, verbose = getOption("Rbitcoin.verbose",
  0))
}
\arguments{
\item{method}{character. For details see \code{blockchain.api.query}, currently supported \code{'Single Address'} and \code{'Single Transaction'}. If \code{method} missing the function will try to guess it based on first param in \dots.}

\item{verbose}{integer. Rbitcoin processing messages, print to console if \code{verbose > 0}, each subfunction reduce \code{verbose} by 1. If missing then \code{getOption("Rbitcoin.verbose",0)} is used.}

\item{\dots}{params passed to blockchain.info API, specific for particular method, example \code{'bitcoin_address'} or \code{'tx_index'}, for more read \code{\link{blockchain.api.query}}.}
}
\value{
post processed response from blockchain api, list or data.table.
}
\description{
Query and process results from blockchain.info.
}
\details{
By default it will perform antiddos check and wait if required, it can be turned off but in such case you should expect to be banned quite easily. Read \code{antiddos}.
}
\examples{
\dontrun{
# Rbitcoin donation address wallet
Rbitcoin_donation_wallet <- blockchain.api.process('15Mb2QcgF3XDMeVn6M7oCG6CQLw4mkedDi')
# some transaction
tx <- blockchain.api.process('e5c4de1c70cb6d60db53410e871e9cab6a0ba75404360bf4cda1b993e58d45f8')
tx
}
}
\references{
\url{https://blockchain.info/api/blockchain_api}
}
\seealso{
\code{\link{blockchain.api.query}}, \code{\link{antiddos}}
}

