% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ct_dict}
\alias{ct.dict}
\alias{ct_dict}
\title{Currency type (fiat or crypti) dictionary}
\usage{
ct_dict()
}
\description{
By default provided to \code{getOption("Rbitcoin.ct.dict")}. Simple crypto/fiat indicator, used in few places, can be extended here.
}

