% Generated by roxygen2: do not edit by hand
\name{fromBTC}
\alias{fromBTC}
\title{Fast convert to/from BTC}
\usage{
fromBTC(value = 1, currency = "USD",
  antiddos = getOption("Rbitcoin.antiddos", TRUE),
  verbose = getOption("Rbitcoin.verbose", 0))
}
\arguments{
\item{value}{numeric.}

\item{currency}{character.}

\item{antiddos}{logical default \code{TRUE}. Default 10s, read \code{\link{antiddos}}.}

\item{verbose}{integer. Rbitcoin processing messages, print to console if \code{verbose > 0}, each subfunction reduce \code{verbose} by 1. If missing then \code{getOption("Rbitcoin.verbose",0)} is used.}
}
\value{
converted value.
}
\description{
Fast convert to/from BTC based on blockchain.info exchange rates API.
}
\examples{
\dontrun{
# 1000 USD (default) to BTC
toBTC(1000)
toBTC(1000, "GBP")
# 1 BTC to USD (default)
fromBTC(1)
fromBTC(1, "EUR")
}
}
\references{
\url{https://blockchain.info/api/exchange_rates_api}
}
\seealso{
\code{\link{toBTC}}, \code{\link{blockchain.api.query}}, \code{\link{antiddos}}
}

