\name{get_rate}
\alias{get_rate}
\title{Get rate}
\usage{
get_rate(v_market, v_base, v_quote, antispam_interval, api.dict,
  verbose = getOption("Rbitcoin.verbose"))
}
\description{
Download ask-bid data for provided market and currency
pair, supports bitcoin markets and yahoo.
}
\keyword{internal}

