% Generated by roxygen2 (4.0.2): do not edit by hand
\name{market.api.query.hitbtc}
\alias{market.api.query.hitbtc}
\title{Send request to hitbtc market API}
\usage{
market.api.query.hitbtc(url, key, secret, req = list(),
  verbose = getOption("Rbitcoin.verbose", 0))
}
\arguments{
\item{url}{character with url on which query needs to be passed.}

\item{key}{character API key used in private API calls.}

\item{secret}{character API secret used in private API calls.}

\item{req}{list of object passed to API: price and amount of opening order, id of cancelling order, etc.}

\item{verbose}{integer. Rbitcoin processing messages, print to console if \code{verbose > 0}, each subfunction reduce \code{verbose} by 1. If missing then \code{getOption("Rbitcoin.verbose",0)} is used.}
}
\value{
Character string a response from market's API call.
}
\description{
Send request to hitbtc market API.
}
\examples{
\dontrun{
# ticker
market.api.query.hitbtc(url = "https://api.hitbtc.com/api/1/public/BTCUSD/ticker")

# wallet
market.api.query.hitbtc(url = "https://api.hitbtc.com/api/1/trading/balance",
                        key = '', secret = '')
}
}
\references{
\url{https://github.com/hitbtc-com/hitbtc-api}
}
\seealso{
\code{\link{market.api.query}}
}
\keyword{internal}

