% Generated by roxygen2: do not edit by hand
\name{market.api.query}
\alias{market.api.query}
\title{Send request to market API}
\usage{
market.api.query(market, url, ..., antiddos = getOption("Rbitcoin.antiddos",
  TRUE), query.dict = getOption("Rbitcoin.query.dict",
  stop("query.dict is mandatory for redirection to market query function")),
  json.debug = getOption("Rbitcoin.json.debug", FALSE),
  verbose = getOption("Rbitcoin.verbose", 0))
}
\arguments{
\item{market}{character which identifies market on which we want to send request, e.g. kraken.}

\item{url}{character url on which launch api call.}

\item{antiddos}{logical default \code{TRUE}. Default to 10s. To customize read \code{antiddos}.}

\item{query.dict}{data.table default built-in dictionary with market query functions, read \code{query.dict}.}

\item{json.debug}{logical default \code{FALSE}. Purely technical, if set to \code{TRUE} it will allow to debug case of unknown non-json format returned by market, by saving market response in working directory as timestamped txt file in case of \code{fromJSON} error.}

\item{verbose}{integer. Rbitcoin processing messages, print to console if \code{verbose > 0}, each subfunction reduce \code{verbose} by 1. If missing then \code{getOption("Rbitcoin.verbose",0)} is used.}

\item{\dots}{objects to be passed to api: \code{key}, \code{secret}, \code{req}, \code{client_id} (used on bitstamp).}
}
\value{
R object created by \code{fromJSON} decoded response from markets API call.
}
\description{
Route a request to particular market function according to dictionary \code{query.dict}.
}
\examples{
\dontrun{
# ticker
market.api.query(market = 'bitstamp',
                 url = 'https://www.bitstamp.net/api/ticker/')
market.api.query(market = 'btce',
                 url = 'https://btc-e.com/api/3/ticker/ltc_usd')
market.api.query(market = 'kraken',
                 url = 'https://api.kraken.com/0/public/Ticker?pair=XXBTZEUR')
market.api.query(market = 'bitmarket',
                 url = 'https://www.bitmarket.pl/json/LTCPLN/ticker.json')
market.api.query(market = 'hitbtc',
                 url = 'https://api.hitbtc.com/api/1/public/BTCUSD/ticker')
# wallet
market.api.query(market = 'bitstamp',
                 url = 'https://www.bitstamp.net/api/balance/',
                 client_id = '', # bitstamp specific
                 key = '', secret = '')
market.api.query(market = 'btce',
                 url = 'https://btc-e.com/tapi',
                 req = list(method = 'getInfo'),
                 key = '', secret = '')
market.api.query(market = 'kraken',
                 url = 'https://api.kraken.com/0/private/Balance',
                 key = '', secret = '')
market.api.query(market = 'bitmarket',
                 url = 'https://www.bitmarket.pl/api2/',
                 req = list(method = 'info'),
                 key = '', secret = '')
market.api.query(market = 'hitbtc',
                 url = "https://api.hitbtc.com/api/1/trading/balance",
                 key = '', secret = '')
# order book
market.api.query(market = 'kraken',
                 url = 'https://api.kraken.com/0/public/Depth?pair=XXBTZEUR')
# open orders
market.api.query(market = 'kraken',
                 url = 'https://api.kraken.com/0/private/OpenOrders',
                 key = '', secret = '')
# place order
market.api.query(market = 'kraken',
                 url = 'https://api.kraken.com/0/private/AddOrder',
                 key = '', secret = '',
                 req = list(pair = 'XXBTZEUR',
                            type = 'sell',
                            ordertype = 'limit',
                            price = 1200, # 1200 eur
                            volume = 0.1)) # 0.1 btc
# cancel order
market.api.query(market = 'kraken',
                 url = 'https://api.kraken.com/0/private/CancelOrder',
                 key = '', secret = '',
                 req = list(txid = 'id_from_open_orders'))
# trades
market.api.query(market = 'kraken',
                 url = 'https://api.kraken.com/0/public/Trades?pair=XXBTZEUR')
}
}
\references{
API documentation: \url{https://www.bitstamp.net/api/}, \url{https://btc-e.com/api/documentation}, \url{https://www.kraken.com/help/api}, \url{https://www.bitmarket.pl/docs.php?file=api_private.html}, \url{https://github.com/hitbtc-com/hitbtc-api}
}
\seealso{
\code{\link{market.api.process}}, \code{\link{antiddos}}, \code{\link{query.dict}}
}

