% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.btc.order_book}
\alias{plot.btc.order_book}
\title{Plot order book data}
\usage{
\method{plot}{btc.order_book}(x,
  limit_pct = getOption("Rbitcoin.plot.limit_pct", Inf),
  verbose = getOption("Rbitcoin.verbose", 0))
}
\arguments{
\item{x}{list, a result from \code{market.api.process(action="order_book")}.}

\item{limit_pct}{numeric, percentage of limit from middle price. It acts like a zoom-in to the middle of order book plot.}
}
\description{
Plot order book data
}
\examples{
\dontrun{
order_book <- market.api.process('bitmarket',c('BTC','PLN'),'order_book')
plot(order_book)
plot(order_book, limit_pct = 0.5)
}
}
\seealso{
\code{\link{market.api.process}}, \code{\link{plot.btc}}
}

