% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.btc}
\alias{Rbitcoin.plot}
\alias{plot.btc}
\title{Plot Rbitcoin objects}
\usage{
\method{plot}{btc}(x, ...)
}
\arguments{
\item{x}{object to be plotted, result of Rbitcoin function, currently supported: \code{market.api.process(action=c("trades","order_book"))}, \code{wallet_manager(archive_read=TRUE)}.}

\item{\dots}{additional params to be passed to particular plot function.}
}
\value{
result of the specific \code{plot} function.
}
\description{
Generic function to plot different objects returned by some Rbitcoin functions. The function is just a wrapper to class specific plots and also placeholder for Rbitcoin plot doc.
}
\note{
Legend may not scale well in interactive viewer, for better scaling save plot to file, see examples.
}
\examples{
\dontrun{
# plot wallet_manager archive to svg
wallet_dt <- wallet_manager(archive_write = FALSE, archive_read = TRUE)
svg()
plot(wallet_dt)
dev.off()
}
}
\seealso{
\code{\link{plot.btc.trades}}, \code{\link{plot.btc.order_book}}, \code{\link{plot.btc.wallet_manager}}
}

