% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rbtc.plot.wallet_manager}
\alias{rbtc.plot.wallet_manager}
\title{Plot wallet manager results}
\usage{
rbtc.plot.wallet_manager(x, y.type = c("value", "amount"),
  mask = getOption("Rbitcoin.plot.mask", FALSE),
  verbose = getOption("Rbitcoin.verbose", 0))
}
\arguments{
\item{x}{data.table, result from \code{wallet_manager(archive_read=TRUE)}}

\item{y.type}{character, measure to be used on Y axis, it will affect a global type of plot. Only "value" currently supported.}

\item{mask}{logical default \code{FALSE}. If set to \code{TRUE} it will mask any amounts and blockchain addresses.}

\item{verbose}{integer. Rbitcoin processing messages, print to console if \code{verbose > 0}, each subfunction reduce \code{verbose} by 1. If missing then \code{getOption("Rbitcoin.verbose",0)} is used.}
}
\value{
\code{TRUE} invisibily. Side effect a base plot - dashboard panel to track different measures of your assets and its value. Use \code{mask} if you want to share the results to somebody, it will overwrite value with value ratio and also mask blockchain addresses.
}
\description{
Plot wallet manager results archive
}
\section{wallet manager archive}{

To be able to track wallet assets value over time user needs to use \code{archive_write=TRUE} at least twice in \code{wallet_manager} processing (with non-NA measures). Using the cryptocurrency which do not have any exchange path to \code{transfer_currency_pair} and/or \code{value_currency} will result \code{NA} as \code{value}. Error on data downloading from external sources (wallets or exchange rates) will also result \code{NA}. Any wallet processing batch which will contains at least one \code{NA} measure will be omitted from plot. If you have some crypto not currenctly supported you may extend dictionary for more currencies or provide its value as manual source to \code{wallet_manager} already calculated in common value currency, remember to comment out the previous source which returns the \code{NA} measure.\cr To plot wallet manager data load wallet archive data, see examples.
Target value currency is taken from the last execution of \code{wallet_manager} archive.
}
\examples{
\dontrun{
wallet_dt <- wallet_manager(archive_write = FALSE, archive_read = TRUE)
rbtc.plot(wallet_dt)
rbtc.plot(wallet_dt, mask=TRUE)
}
}
\seealso{
\code{\link{wallet_manager}}, \code{\link{rbtc.plot}}
}

