% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setclass}
\alias{setclass}
\title{setclass}
\usage{
setclass(x, newclass, append = TRUE)
}
\description{
setclass by reference, put new class as the first element
}
\references{
https://github.com/Rdatatable/data.table/pull/844
}
\keyword{internal}

