% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toBTC}
\alias{toBTC}
\title{Fast convert to/from BTC}
\usage{
toBTC(currency = "USD", value = 1)
}
\arguments{
\item{currency}{character.}

\item{value}{numeric.}
}
\value{
converted value.
}
\description{
Fast convert to/from BTC based on blockchain.info exchange rates API.
}
\examples{
\dontrun{
# 1000 USD to BTC
toBTC("USD", 1000)
# 1 BTC to USD
fromBTC("USD", 1)
}
}
\references{
\url{https://blockchain.info/api/exchange_rates_api}
}
\seealso{
\code{\link{blockchain.api.query}}
}

