% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_rate}
\alias{get_rate}
\title{Get rate}
\usage{
get_rate(v_market, v_base, v_quote, api.dict,
  antiddos = getOption("Rbitcoin.antiddos", TRUE),
  verbose = getOption("Rbitcoin.verbose", 0))
}
\description{
Download ask-bid data for provided market and currency pair, supports bitcoin markets and yahoo.
}
\keyword{internal}

