% Generated by roxygen2: do not edit by hand
\name{query_dict}
\alias{query.dict}
\alias{query_dict}
\title{Market API query function dictionary}
\usage{
query_dict()
}
\description{
By default provided to \code{getOption("Rbitcoin.query.dict")}. This function returns built-in Rbitcoin supported markets dictionary. Granularity of query dictionary data is \code{c("market")}. This dictionary can be edited/extended by user for new markets.\cr Currently supported:
\itemize{
\item \code{kraken}
\item \code{bitstamp}
\item \code{btce}
\item \code{bitmarket}
\item \code{hitbtc}
\item \code{btcchina}
}
}
\note{
Do not use \code{query.dict} from untrusted source or read whole it's code to ensure it is safe! The api dictionary was not fully tested, please follow the examples, if you find any bugs please report.
}
\examples{
\dontrun{
query_dict()[]
}
}

