% Generated by roxygen2: do not edit by hand
\name{rbtc.plot.trades}
\alias{rbtc.plot.trades}
\title{Plot trades data}
\usage{
rbtc.plot.trades(x, ..., verbose = getOption("Rbitcoin.verbose", 0))
}
\arguments{
\item{x}{list, a result from \code{market.api.process(action="trades")}.}

\item{verbose}{integer. Rbitcoin processing messages, print to console if \code{verbose > 0}, each subfunction reduce \code{verbose} by 1. If missing then \code{getOption("Rbitcoin.verbose",0)} is used.}

\item{\dots}{additional params to be passed to particular plot function.}
}
\description{
Plot trades data
}
\examples{
\dontrun{
trades <- market.api.process('kraken',c('BTC','EUR'),'trades')
rbtc.plot(trades)
}
}
\seealso{
\code{\link{market.api.process}}, \code{\link{rbtc.plot}}
}

