% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rbtc.plot.wallet_manager}
\alias{rbtc.plot.wallet_manager}
\title{Plot wallet manager results}
\usage{
rbtc.plot.wallet_manager(wallet_dt, type = c("value", "recent"),
  mask = getOption("Rbitcoin.plot.mask", FALSE),
  verbose = getOption("Rbitcoin.verbose", 0))
}
\arguments{
\item{wallet_dt}{data.table, result from \code{wallet_manager}}

\item{type}{character, type of plot: \code{"value"} - wallet value in time, \code{"amount"} - wallet amount in time (not yet supported), \code{"recent"} - recent wallet.}

\item{mask}{logical default \code{FALSE}. If set to \code{TRUE} it will mask sensitive informations such as: amounts, values, blockchain addresses, auth accounts.}

\item{verbose}{integer. Rbitcoin processing messages, print to console if \code{verbose > 0}, each subfunction reduce \code{verbose} by 1. If missing then \code{getOption("Rbitcoin.verbose",0)} is used.}
}
\value{
invisibily \code{NULL}. Side effect a base plot, for type \code{"value"} or \code{"amount"} the dashboard panel to track different measures of your assets over time. For type \code{"recent"} a base plot of the most recent non-NA measures wallet data.
}
\description{
Plot wallet manager results archive
}
\section{wallet manager archive}{

To be able to track wallet assets value over time user needs to use \code{archive_write=TRUE} at least twice in \code{wallet_manager} processing (with non-NA measures). Using the cryptocurrency which do not have any exchange path to \code{transfer_currency_pair} and/or \code{value_currency} will result \code{NA} as \code{value}. Error on data downloading from external sources (wallets or exchange rates) will also result \code{NA}. Any wallet processing batch which will contains at least one \code{NA} measure will be omitted from plot. If you have some crypto not currenctly supported you may extend dictionary for more currencies or provide its value as manual source to \code{wallet_manager} already calculated in common value currency, remember to comment out the previous source which returns the \code{NA} measure.\cr To plot wallet manager data load wallet archive data, see examples.
Target value currency is taken from the last execution of \code{wallet_manager} archive.
}
\examples{
\dontrun{
wallet_dt <- wallet_manager(archive_write=FALSE, archive_read=TRUE)
# plot wallet value over time
rbtc.plot(wallet_dt)
# the same but with masked sensitive data
rbtc.plot(wallet_dt, mask=TRUE)
# legend lookS better when plot to file
svg("wallet_value.svg")
rbtc.plot(wallet_dt, type="value", mask=TRUE)
dev.off()
# plot recent wallet value
rbtc.plot(wallet_dt, type="recent")
}
}
\seealso{
\code{\link{wallet_manager}}, \code{\link{rbtc.plot}}
}

